cos/admission.rs: colocate 8 tests with production (#984 P3 Phase 3b)
Move 8 tests from tx/mod.rs::tests to cos/admission.rs::tests covering
cos_flow_aware_buffer_limit (7 tests) and cos_queue_flow_share_limit (1).

Adds COS_FLOW_FAIR_BUCKETS, CoSQueueConfig imports.

cargo test --bins 865/0/2.
